+++ console/src/components/Services/VoyagerView/VoyagerView.tsx
@@ -1,18 +1,15 @@
-import React, { Component } from 'react';
-import { Connect } from 'react-redux';
+import React from 'react';
+import { connect } from 'react-redux';
 import { GraphQLVoyager } from 'graphql-voyager';
 
 import Endpoints from '../../../Endpoints';
+import ErrorBoundary from '../Common/ErrorBoundary';
+import { Dispatch, ReduxState } from '../../../types';
 import '../../../../node_modules/graphql-voyager/dist/voyager.css';
 import './voyagerView.css';
 import requestAction from '../../../utils/requestAction';
-import { Dispatch } from '../../../types';
 
-interface VoyagerViewProps {
-  headers: Headers;
-}
-
-const mapStateToProps = (state: State) => {
+const mapStateToProps = (state: ReduxState) => {
   return {
     headers: state.tables.dataHeaders,
   };
@@ -23,37 +20,49 @@ const mapDispatchToProps = (dispatch: Dispatch) => {
       dispatch(requestAction(url, options)),
   };
 };
-// TODO: replace by redux State when it's defined
-interface State {
-  tables: {
-    dataHeaders: Headers;
-  };
-}
 
-type Props = VoyagerViewProps &
-  ReturnType<typeof mapStateToProps> &
-  ReturnType<typeof mapDispatchToProps>;
+interface VoyagerViewProps
+  extends ReturnType<typeof mapStateToProps>,
+    ReturnType<typeof mapDispatchToProps> {}
 
-class VoyagerView extends Component<Props, State> {
-  introspectionProvider = (query: string) =>
-    this.props.requestAction(Endpoints.graphQLUrl, {
+const VoyagerView = (props: VoyagerViewProps) => {
+  const introspectionProvider = (query: string) =>
+    props.requestAction(Endpoints.graphQLUrl, {
       method: 'POST',
-      headers: this.props.headers,
+      headers: props.headers,
       body: JSON.stringify({ query }),
     });
 
-  render() {
-    return (
-      <GraphQLVoyager
-        introspection={this.introspectionProvider}
-        workerURI="https://cdn.jsdelivr.net/npm/graphql-voyager@1.0.0-rc.27/dist/voyager.worker.min.js"
-      />
-    );
-  }
-}
+  const loadWorker = async () => {
+    const url = Endpoints.voyagerWorker;
+    const response = await fetch(url);
+    const payload = await response.text();
+    // HACK: to increase viz.js memory size from 16mb to 128mb
+    // https://github.com/APIs-guru/graphql-voyager/blob/v1.0.0-rc.29/src/utils/index.ts#L21
+    const newPayload = payload
+      .replace('||16777216;', '||(16777216 * 8);')
+      .replace('||5242880;', '||(5242880 * 8);');
+    const script = new Blob([newPayload], { type: 'application/javascript' });
+    const workerURL = URL.createObjectURL(script);
+    return new Worker(workerURL);
+  };
 
-const generatedVoyagerConnector = (connect: Connect) => {
-  return connect(mapStateToProps, mapDispatchToProps)(VoyagerView);
+  return (
+    <ErrorBoundary
+      message="You might be seeing this because your schema is too large to be handled by Voyager."
+      helmetTitle="Voyager Error | Hasura"
+    >
+      <GraphQLVoyager
+        introspection={introspectionProvider}
+        loadWorker={loadWorker}
+      />
+    </ErrorBoundary>
+  );
 };
 
-export default generatedVoyagerConnector;
+const ConnectedVoyagerView = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(VoyagerView);
+
+export default ConnectedVoyagerView;