+++ CHANGELOG.md
@@ -10,9 +10,11 @@
 - server: support only a bounded plan cache, with a default size of 4000 (closes #5363)
 - server: add logs for action handlers
 - server: add request/response sizes in event triggers (and scheduled trigger) logs
+- console: update sidebar icons for different action and trigger types
 - console: handle nested fragments in allowed queries (close #5137) (#5252)
 - console: update sidebar icons for different action and trigger types (#5445)
 - console: make add column UX consistent with others (#5486)
+- console: add ErrorBoundary for VoyagerView and fix memory requirements of viz.js (close #4818) (#4856)
 
 ## `v1.3.0`